+++ src/main/java/com/kurkus/kusinsa/dto/request/product/status/ProductStockUpdateRequest.java
@@ -1,0 +1,2 @@
+package com.kurkus.kusinsa.dto.request.product;public class ProductStockUpdateRequest {
+}
